refactor: extract module
Extract editor-version related logic to own module
